requirements.txt
dbus-python == 1.2.18
faiss-cpu
sentence-transformers
numpy
